enable experimentalFeatures.canonicalURLRedirect by default
experimentalFeatures.canonicalURLRedirect was introduced in Sourcegraph 2.9 and was disabled by default. We have been running with it enabled in dev for 4 months. In 2.13 it should be enabled by default so that we can remove the feature flag in the following month and eliminate tech debt.

Also filed https://github.com/sourcegraph/sourcegraph/issues/466 to schedule the removal of this feature flag.